+++ lib/tty/progressbar/formatter/estimated.rb
@@ -22,7 +22,7 @@ module TTY
           return value.gsub(matcher, "--s")
         end
 
-        elapsed   = Time.now - @progress.start_at
+        elapsed = @progress.elapsed_time
         estimated = (elapsed / @progress.ratio).to_f - elapsed
         estimated = (estimated.infinite? || estimated < 0) ? 0.0 : estimated
         value.gsub(matcher, Converter.to_time(estimated))